JMS and Spring Message Data Types
javax.jms.Message is an interface for JMS-implementation; org.springframework.messaging.Message is the Spring framework abstraction of it - if an application is developed with Spring Boot, it's better to use the latter, because switching between Brokern is easier.

With this commit and a Breakpoint at 'listenForHello(...)' the additional header informations in the Spring framework "Message" can be compared to the ones in the javax.jms.Message.

https://www.udemy.com/course/spring-framework-5-beginner-to-guru/learn/lecture/17022484, Lektion 29-462